chore: moved the add function to services and updated args
